bench: move userspace suite into benchmark/ with build tag
Drop the separate benchmark-userspace/ directory. Instead, add the
userspace benchmarks as bench_userspace_test.go in the existing
benchmark/ package, gated by the 'userspace' build tag.

The kernel and userspace benchmarks can't share a TestMain (bpftime's
syscall-server intercepts BPF calls, which would break kernel BPF).
Build tags solve this cleanly: only one TestMain is compiled at a time.

  make bench           # kernel uprobe suite (sudo required)
  make bench-kernel    # same as bench
  make bench-userspace # userspace BPF suite (no sudo, needs bpftime-libs)

Reports: bench-report.json (kernel) and bench-report-userspace.json.
