+++ benchmark/bench_test.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build linux && !userspace
 
 package benchmark
